feat: support round-robin with multiple engines (#720)
Preliminary work for supporting round-robins with multiple engines, output looks like the following:

```
Rank Name                             Elo        +/-       nElo        +/-      Games      Score       Draw           Ptnml(0-2)
   1 engine2                        34.86      67.31      65.66     124.33         30      55.0%      60.0%      [0, 2, 9, 3, 1]
   2 engine1                        -0.00      56.08       0.00     124.33         30      50.0%      80.0%     [1, 0, 12, 2, 0]
   3 engine3                       -11.59      39.08     -37.04     124.33         30      48.3%      80.0%     [0, 2, 12, 1, 0]
   4 engine4                       -23.20      44.30     -65.66     124.33         30      46.7%      73.3%     [0, 3, 11, 1, 0]

```

+++ app/src/matchmaking/tournament/base/tournament.hpp
@@ -69,12 +69,12 @@ class BaseTournament {
     std::uint64_t initial_matchcount_;
 
    private:
-    std::uint64_t setResults(const stats_map &results) noexcept {
+    std::size_t setResults(const stats_map &results) noexcept {
         Logger::trace("Setting results...");
 
         scoreboard_.setResults(results);
 
-        std::uint64_t total = 0;
+        std::size_t total = 0;
 
         for (const auto &pair1 : scoreboard_.getResults()) {
             const auto &stats = pair1.second;
